big integer para usuario

+++ src/database/migrations/2019_02_21_174355_create_rbac_table.php
@@ -33,7 +33,7 @@ class CreateRbacTable extends Migration
         });
 
         Schema::create('role_user', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->unsignedInteger('role_id');
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
@@ -41,7 +41,7 @@ class CreateRbacTable extends Migration
         });
 
         Schema::create('permission_user', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->unsignedInteger('permission_id');
             $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
